finally implemented that suggestion from AndzCLiv3

+++ lib/cogs/misc.py
@@ -43,13 +43,17 @@ class Misc(Cog):
 	@cooldown(1, 4, BucketType.user)
 	async def time_bomb_command(self, ctx, time: int, *, message : str):
 		"""Makes a message with a timelimit on it, the time is in seconds."""
-		await ctx.message.delete()
+		if time >= 10000:
+			await ctx.message.delete()
 
-		embed = Embed(title="Time Message", description=message, colour=ctx.author.colour)
-		embed.set_thumbnail(url=ctx.author.avatar_url)
-		embed.set_footer(text=f"{ctx.author} | This message only lasts {str(time)} seconds.", icon_url=ctx.author.avatar_url)
+			embed = Embed(title="Time Message", description=message, colour=ctx.author.colour)
+			embed.set_thumbnail(url=ctx.author.avatar_url)
+			embed.set_footer(text=f"{ctx.author} | This message only lasts {str(time)} seconds.", icon_url=ctx.author.avatar_url)
 
-		await ctx.send(embed=embed, delete_after=time)
+			await ctx.send(embed=embed, delete_after=time)
+
+		else:
+			await ctx.send("Please try again with a lower time.", delete_after=15)
 
 	@Cog.listener()
 	async def on_ready(self):
